auth/requirements.txt
cryptography==42.0.5
fastapi==0.115.5
nest_asyncio==1.6.0
passlib==1.7.4
pydantic==2.10.2
PyJWT==2.10.1
uvicorn==0.32.1
